refactor: split the ticket thread store
stores/ticket/ now holds thread, details, feedback and the composer as
siblings; the chat/timeline bubble geometry — presentation data the
blocks bind — moves beside the ticket components as bubbleLayout.ts.
The thread module keeps its useTicketThread facade and spreads the
composer as before.
